requirements.txt
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain
pandas
requests
ipykernel
python-dotenv
beautifulsoup4
openai
langchain-openai
langchain-community
unstructured
duckduckgo-search
matplotlib
seaborn
openpyxl